refactor of components placement in the file structure
