+++ apps/client/src/components/common/MapPreview.tsx
