+++ apps/client/src/components/index.ts
@@ -1,13 +1,13 @@
-export * from "./RequireAuth";
-export * from "./DashboardSwitch";
+export * from "./layers/RequireAuth.tsx";
+export * from "./layers/DashboardSwitch.tsx";
 export * from "./office";
 export * from "./patients";
 export * from "./appointments";
 export * from "./SideMenu";
-export * from "./LanguageSwitcher.tsx";
-export * from "./SubmitButton/SubmitButton.tsx";
+export * from "./common/LanguageSwitcher.tsx";
+export * from "./common/SubmitButton/SubmitButton.tsx";
 export * from "./AuthLinks/AuthFooter.tsx";
 export * from "./RoleGuard/RoleGuard.tsx";
 export * from "./DataTable/DataTable.tsx";
-export * from "./GoogleAuthButton/GoogleAuthButton.tsx";
-export * from "./PhoneInputField/PhoneInputField.tsx";
+export * from "./Login/GoogleAuthButton.tsx";
+export * from "./common/PhoneInputField.tsx";
