+++ apps/client/src/components/layers/DashboardSwitch.tsx
@@ -1,5 +1,5 @@
-import { USER_ROLES } from "../types/user";
-import { useUser } from "../hooks/useUser.ts";
+import { USER_ROLES } from "../../types/user.ts";
+import { useUser } from "../../hooks/useUser.ts";
 import { JSX } from "react";
 import { Navigate } from "react-router";
 import { ROUTER_PATHS } from "@router";
